scripts/resize_annotated_requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ijson==3.3.0
joblib==1.4.2
jsonschema==3.2.0
label-studio-sdk==1.0.3
lxml==5.2.2
nltk==3.6.7
numpy==2.0.0
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.0
pydantic_core==2.20.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
requests==2.32.3
requests-mock==1.12.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
xmljson==0.2.1
